test: minimal repro for cloudflare rewrites issue
Reproduces an issue where the request URL for a rewritten request is different
based on whether the app is running locally or on Vercel vs. Cloudflare. On
local and Vercel, the request URL is that of the path prior to the rewrite.
On Cloudflare, the request URL is that of rewrite destination.